Moved MainActivity to activities->navigation package
